uth_project_fastapi_extracted/requirements.txt
fastapi
uvicorn[standard]
pydantic
motor
python-dotenv
bson
bcrypt
python-jose[cryptography]
passlib[bcrypt]
